% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_ga.R
\name{add_update_params}
\alias{add_update_params}
\title{add_update_params}
\usage{
add_update_params(org, new)
}
\arguments{
\item{org}{a original list}

\item{new}{a new list}
}
\description{
add_update_params
}
