% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_ga.R
\name{fit_ga_search}
\alias{fit_ga_search}
\title{fit_ga_search}
\usage{
fit_ga_search(id, data, model, params, static, ...)
}
\arguments{
\item{params}{a list of dial parameter objects}

\item{static}{a list of static data}
}
\description{
fit_ga_search
}
