% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_cnn_lstm.R
\name{keras_cnn_lstm}
\alias{keras_cnn_lstm}
\title{keras cnn lstm}
\usage{
keras_cnn_lstm(input_dim, embed_dim = 128, seq_len = 50,
  filter_size = 5, n_filters = 100, pool_size = 4, lstm_dim = 64,
  lstm_drop = 0.2, bidirectional = F, dropout = 0.2,
  output_dim = 2, output_fun = "softmax")
}
\arguments{
\item{input_dim}{Number of unique vocabluary/tokens}

\item{embed_dim}{Number of word vectors}

\item{seq_len}{Length of the input sequences}

\item{filter_size}{the window size (kernel_size)}

\item{n_filters}{the number of convolutional filters}

\item{pool_size}{pooling dimension (filters)}

\item{lstm_dim}{Number of lstm neurons (default 32)}

\item{lstm_drop}{default is 2}

\item{bidirectional}{default is F}

\item{output_dim}{Number of neurons of the output layer}

\item{output_fun}{Output activation function}
}
\value{
keras model
}
\description{
Word embedding + 1D pooled convolution + lstm layer
}
