% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_gru_cnn.R
\name{keras_gru_cnn}
\alias{keras_gru_cnn}
\title{keras gru cnn}
\usage{
keras_gru_cnn(input_dim, embed_dim = 128, seq_len = 50, gru_dim = 64,
  gru_drop = 0.2, filter_sizes = c(3, 2), n_filters = c(120, 60),
  pool_size = 4, output_fun = "softmax", output_dim = 1)
}
\arguments{
\item{input_dim}{Number of unique vocabluary/tokens}

\item{embed_dim}{Number of word vectors}

\item{seq_len}{Length of the input sequences}

\item{gru_dim}{Number of lstm neurons (default 32)}

\item{gru_drop}{default is 2}

\item{n_filters}{the number of convolutional filters}

\item{pool_size}{pooling dimension (filters)}

\item{output_fun}{Output activation function}

\item{output_dim}{Number of neurons of the output layer}

\item{filter_size}{the window size (kernel_size)}
}
\value{
keras model
}
\description{
Word embedding + gru global average & max + 1D pooled convolution
}
