% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_multi_cnn.R
\name{keras_multi_cnn}
\alias{keras_multi_cnn}
\title{keras cnn gru}
\usage{
keras_multi_cnn(input_dim, embed_dim = 128, seq_len = 50,
  filter_sizes = c(1, 2, 3, 4), num_filters = 50, output_dim = 2,
  output_fun = "softmax")
}
\arguments{
\item{input_dim}{Number of unique vocabluary/tokens}

\item{embed_dim}{Number of word vectors}

\item{seq_len}{Length of the input sequences}

\item{output_dim}{Number of neurons of the output layer}

\item{output_fun}{Output activation function}

\item{n_filters}{the number of convolutional filters}

\item{filter_size}{the window size (kernel_size)}
}
\value{
keras model
}
\description{
Word embedding + 1D pooled convolution + gru layer
}
\details{
pool_size is determined automatically
}
