% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_simple_lstm.R
\name{keras_simple_lstm}
\alias{keras_simple_lstm}
\title{keras simple lstm}
\usage{
keras_simple_lstm(input_dim, embed_dim = 128, seq_len = 50,
  lstm_dim = 64, lstm_drop = 0.2, dropout = 0.2, bidirectional = F,
  output_dim = 2, output_fun = "softmax")
}
\arguments{
\item{input_dim}{Number of unique vocabluary/tokens}

\item{embed_dim}{Number of word vectors}

\item{seq_len}{Length of the input sequences}

\item{lstm_dim}{Number of recurrent neurons (default 64)}

\item{lstm_drop}{Recurrent dropout ratio}

\item{bidirectional}{default is F}

\item{output_dim}{Number of neurons of the output layer}

\item{output_fun}{Output activation function}
}
\value{
keras model
}
\description{
Word embedding + (bidirectional) long short-term memory
}
