% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_simple_mlp.R
\name{keras_simple_mlp}
\alias{keras_simple_mlp}
\title{keras simple mlp}
\usage{
keras_simple_mlp(input_dim, embed_dim, seq_len, pooling = "flatten",
  dense_dim = 128, dense_fun = "relu", dropout = 0.5,
  output_fun = "softmax", output_dim)
}
\arguments{
\item{input_dim}{Number of unique vocabluary/tokens}

\item{embed_dim}{Number of word vectors}

\item{seq_len}{Length of the input sequences}

\item{pooling}{one of c("flatten", "global_average", "average")}

\item{output_fun}{Output activation function}

\item{output_dim}{Number of neurons of the output layer}
}
\value{
keras model

\if{html}{
\out{
 <img src = "https://media.giphy.com/media/IB9foBA4PVkKA/giphy.gif">
}}
}
\description{
Word Embedding + Simple Multilayer Perceptron
}
