% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrMBO.R
\name{run_mbo}
\alias{run_mbo}
\title{run_mbo}
\usage{
run_mbo(data, params, const = NULL, prior = NULL, n_init = 5,
  n_main = 30, name = "", metric = "accuracy", parallel = F)
}
\arguments{
\item{data}{input}

\item{params}{params}

\item{const}{constants}

\item{n_init}{burn in iterations}

\item{n_main}{main iterations}
}
\value{
list(data = data, params = params)
}
\description{
tidy MBO
}
